# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/organizations/{organizationId}/reports/sales": {
      "get": {
        "operationId": "ReportsController_getSalesReport",
        "parameters": [
          {
            "name": "eventId",
            "required": false,
            "in": "query",
            "description": "Filter nach Event-ID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Startdatum für den Bericht",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Enddatum für den Bericht",
            "schema": {
              "example": "2024-12-31T23:59:59.000Z",
              "type": "string"
            }
          },
          {
            "name": "groupBy",
            "required": false,
            "in": "query",
            "description": "Gruppierung der Berichtsdaten",
            "schema": {
              "example": "day",
              "type": "string",
              "enum": [
                "day",
                "week",
                "month",
                "category",
                "product",
                "payment_method"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/organizations/{organizationId}/reports/products": {
      "get": {
        "operationId": "ReportsController_getProductsReport",
        "parameters": [
          {
            "name": "eventId",
            "required": false,
            "in": "query",
            "description": "Filter nach Event-ID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Startdatum für den Bericht",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Enddatum für den Bericht",
            "schema": {
              "example": "2024-12-31T23:59:59.000Z",
              "type": "string"
            }
          },
          {
            "name": "groupBy",
            "required": false,
            "in": "query",
            "description": "Gruppierung der Berichtsdaten",
            "schema": {
              "example": "day",
              "type": "string",
              "enum": [
                "day",
                "week",
                "month",
                "category",
                "product",
                "payment_method"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/organizations/{organizationId}/reports/payments": {
      "get": {
        "operationId": "ReportsController_getPaymentsReport",
        "parameters": [
          {
            "name": "eventId",
            "required": false,
            "in": "query",
            "description": "Filter nach Event-ID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Startdatum für den Bericht",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Enddatum für den Bericht",
            "schema": {
              "example": "2024-12-31T23:59:59.000Z",
              "type": "string"
            }
          },
          {
            "name": "groupBy",
            "required": false,
            "in": "query",
            "description": "Gruppierung der Berichtsdaten",
            "schema": {
              "example": "day",
              "type": "string",
              "enum": [
                "day",
                "week",
                "month",
                "category",
                "product",
                "payment_method"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/organizations/{organizationId}/reports/hourly": {
      "get": {
        "operationId": "ReportsController_getHourlyReport",
        "parameters": [
          {
            "name": "eventId",
            "required": false,
            "in": "query",
            "description": "Filter nach Event-ID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Startdatum für den Bericht",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Enddatum für den Bericht",
            "schema": {
              "example": "2024-12-31T23:59:59.000Z",
              "type": "string"
            }
          },
          {
            "name": "groupBy",
            "required": false,
            "in": "query",
            "description": "Gruppierung der Berichtsdaten",
            "schema": {
              "example": "day",
              "type": "string",
              "enum": [
                "day",
                "week",
                "month",
                "category",
                "product",
                "payment_method"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/organizations/{organizationId}/reports/inventory": {
      "get": {
        "operationId": "ReportsController_getInventoryReport",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/organizations/{organizationId}/reports/stock-movements": {
      "get": {
        "operationId": "ReportsController_getStockMovementsReport",
        "parameters": [
          {
            "name": "eventId",
            "required": false,
            "in": "query",
            "description": "Filter nach Event-ID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Startdatum für den Bericht",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Enddatum für den Bericht",
            "schema": {
              "example": "2024-12-31T23:59:59.000Z",
              "type": "string"
            }
          },
          {
            "name": "groupBy",
            "required": false,
            "in": "query",
            "description": "Gruppierung der Berichtsdaten",
            "schema": {
              "example": "day",
              "type": "string",
              "enum": [
                "day",
                "week",
                "month",
                "category",
                "product",
                "payment_method"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/organizations/{organizationId}/reports/export": {
      "get": {
        "operationId": "ReportsController_exportReport",
        "parameters": [
          {
            "name": "eventId",
            "required": false,
            "in": "query",
            "description": "Filter nach Event-ID",
            "schema": {
              "example": "550e8400-e29b-41d4-a716-446655440000",
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Startdatum für den Bericht",
            "schema": {
              "example": "2024-01-01T00:00:00.000Z",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "Enddatum für den Bericht",
            "schema": {
              "example": "2024-12-31T23:59:59.000Z",
              "type": "string"
            }
          },
          {
            "name": "groupBy",
            "required": false,
            "in": "query",
            "description": "Gruppierung der Berichtsdaten",
            "schema": {
              "example": "day",
              "type": "string",
              "enum": [
                "day",
                "week",
                "month",
                "category",
                "product",
                "payment_method"
              ]
            }
          },
          {
            "name": "format",
            "required": true,
            "in": "query",
            "description": "Exportformat des Berichts",
            "schema": {
              "example": "csv",
              "type": "string",
              "enum": [
                "csv",
                "excel",
                "json"
              ]
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "tags": [
          "Reports"
        ]
      }
    },
...
    "/api/setup": {
      "post": {
        "operationId": "SetupController_performSetup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Einrichtung erfolgreich",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "firstName": {
                          "type": "string"
                        },
                        "lastName": {
                          "type": "string"
                        }
                      }
                    },
                    "organization": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "slug": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Einrichtung bereits abgeschlossen"
          }
        },
        "summary": "Führt die erstmalige Einrichtung durch",
        "tags": [
          "Setup"
        ]
      }
    }
  },
  "info": {
    "title": "OpenEOS API",
    "description": "\n# OpenEOS API Documentation\n\nOpen Event Ordering System - REST API für Kassensysteme bei Vereinen und Veranstaltungen.\n\n## Authentifizierung\n\nDie API verwendet JWT Bearer Tokens für die Authentifizierung. Nach dem Login erhältst du einen Access Token (30 Min gültig) und einen Refresh Token (7 Tage, als httpOnly Cookie).\n\n### Endpoints ohne Authentifizierung\n- `POST /api/auth/register` - Registrierung\n- `POST /api/auth/login` - Login\n- `POST /api/auth/refresh` - Token erneuern\n- `POST /api/auth/forgot-password` - Passwort vergessen\n- `POST /api/auth/reset-password` - Passwort zurücksetzen\n- `GET /api/health/*` - Health Checks\n- `GET /api/invitations/:token` - Einladung abrufen\n- `/api/public/order/*` - Online-Bestellungen (mit Session-Token)\n\n## Multi-Tenant\n\nDie API ist multi-tenant-fähig. Die meisten Endpoints erfordern eine `organizationId` im Pfad.\n\n## Rollen\n\n- **ADMIN** - Vollzugriff auf Organisation\n- **MANAGER** - Verwaltung von Produkten, Events, Bestellungen\n- **CASHIER** - Kassen-Operationen\n- **KITCHEN** - Küchen-Display\n- **DELIVERY** - Ausgabe-Display\n\n## WebSocket\n\nReal-time Updates über Socket.io unter `/` mit Namespaces:\n- `/pos` - Kassen-Clients\n- `/display` - Küchen/Ausgabe-Displays\n- `/printer` - Drucker-Agenten\n- `/admin` - Admin-Dashboard\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "OpenEOS",
...
    },
    "license": {
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "X-Organization-Id": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Organization-Id",
        "description": "Organization ID for multi-tenant requests"
      },
      "X-Session-Token": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Session-Token",
        "description": "Session token for online orders"
      }
    },
    "schemas": {
...
      "SetupDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@example.com",
            "description": "E-Mail-Adresse des Super-Admins"
          },
          "password": {
            "type": "string",
            "example": "SecurePass123!",
            "description": "Passwort (min. 8 Zeichen, Groß-/Kleinbuchstabe und Zahl)",
            "minLength": 8,
            "maxLength": 72
          },
          "firstName": {
            "type": "string",
            "example": "Max",
            "description": "Vorname des Admins",
            "minLength": 2,
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "example": "Mustermann",
            "description": "Nachname des Admins",
            "minLength": 2,
            "maxLength": 100
          },
          "organizationName": {
            "type": "string",
            "example": "Mein Verein e.V.",
            "description": "Name der ersten Organisation",
            "minLength": 2,
            "maxLength": 200
          }
        },
        "required": [
          "email",
          "password",
          "firstName",
          "lastName",
          "organizationName"
        ]
      }
    }
  }
}